222356K_og/requirements.txt
# General Packages
pandas>=2.2.3
PyYAML>=6.0.2
notebook>=7.2.2

# Machine Learning Packages
scikit-learn>=1.5.2
joblib>=1.4.2